Extract IntervalUtils from DateTimeUtils
