+++ h2/src/main/org/h2/command/Parser.java
@@ -190,6 +190,7 @@ import org.h2.table.TableFilter.TableFilterVisitor;
 import org.h2.table.TableView;
 import org.h2.util.DateTimeFunctions;
 import org.h2.util.DateTimeUtils;
+import org.h2.util.IntervalUtils;
 import org.h2.util.MathUtils;
 import org.h2.util.ParserUtil;
 import org.h2.util.StatementBuilder;
@@ -3691,7 +3692,7 @@ public class Parser {
             qualifier = IntervalQualifier.SECOND;
         }
         try {
-            return ValueExpression.get(DateTimeUtils.parseInterval(qualifier, negative, s));
+            return ValueExpression.get(IntervalUtils.parseInterval(qualifier, negative, s));
         } catch (Exception e) {
             throw DbException.get(ErrorCode.INVALID_DATETIME_CONSTANT_2, e, "INTERVAL", s);
         }
